feat: 좌석 예약 요청 구현

+++ src/main/java/hhplus/concert/infra/persistence/ConcertSeatJpaRepository.java
@@ -3,5 +3,5 @@ package hhplus.concert.infra.persistence;
 import hhplus.concert.domain.ConcertSeat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConcertSeatJpaRepository extends JpaRepository<ConcertSeat, Integer> {
+public interface ConcertSeatJpaRepository extends JpaRepository<ConcertSeat, Long> {
 }
